requirements.txt
requests
pytz
influxdb
paho-mqtt
argparse
python-dotenv
ConfigParser
pyserial
dataclasses
jsons
pymysql
Pillow
psutil
